:recycle: move ES related objects out of event folder
