+++ marketplace-domain/src/contribution/mod.rs
@@ -5,7 +5,7 @@ mod event;
 pub use event::Event;
 
 mod aggregate;
-pub use aggregate::{Contribution, Id};
+pub use aggregate::{Aggregate, Id};
 
 mod state;
 pub use state::State;
