+++ marketplace-domain/src/event_aggregate.rs
